Move non endpoint stuff to ``.Functions.cs`
